Dissolve TelPrinter into Showable; make the TEL parser non-public
`TelPrinter` and the `Tel.show` methods were a second, public way to
render TEL, parallel to what a `Showable` instance should provide. The
line-based serialization now lives inside `given Tel.Document is Showable`,
with `given Tel is Showable` wrapping a Compound-rooted Tel in a Document
and composing with it — so `.show` is the single route to TEL text. The
HTTP Postable/Servable and the tests move to `.show`.

The stateful parser stays a class but is no longer a public second entry
point: `object TelParser` and the `Tel.parse`/`parseAll`/`parseStream`
facades become `private[stratiform]`, so external callers must go through
`bytes.read[Tel]` / `text.load[Tel]`. The unused streaming `emit` variant
is dropped.
